#8 Modify to seperate chat and comment ui to fragment
