+++ app/src/main/java/com/song2/thenaun/ui/detailed/chat/ChatFriendItem.kt
@@ -1,12 +1,9 @@
-package com.song2.thenaun.ui.detailed
+package com.song2.thenaun.ui.detailed.chat
 
 import com.song2.thenaun.base.recyclerview.BaseItem
 
-data class ChatItem(
+data class ChatFriendItem(
     override val itemId: String,
-    val type: Boolean,
-    val time: String,
     val profile: Int,
     val nick: String,
-    val contents: String
 ) : BaseItem
