+++ app/src/main/java/com/song2/thenaun/ui/detailed/chat/ChatItem.kt
@@ -1,9 +1,12 @@
-package com.song2.thenaun.ui.detailed
+package com.song2.thenaun.ui.detailed.chat
 
 import com.song2.thenaun.base.recyclerview.BaseItem
 
-data class ChatFriendItem(
+data class ChatItem(
     override val itemId: String,
+    val type: Boolean,
+    val time: String,
     val profile: Int,
     val nick: String,
+    val contents: String
 ) : BaseItem
