+++ app/src/main/res/layout/item_chat_me.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="item"
-            type="com.song2.thenaun.ui.detailed.ChatItem" />
+            type="com.song2.thenaun.ui.detailed.chat.ChatItem" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
